+++ DAL/Models/AvailableDayGetAll.cs
@@ -11,15 +11,15 @@ namespace Furlough.DAL.Models
         public string EmployeeDepartment { get; set; }
         [Display(Name = "Position")]
         public string EmployeePosition { get; set; }
-        public int Medical { get; set; }
-        public int Yearly { get; set; }
-        public int Overtime { get; set; }
-        public int Birth { get; set; }
-        public int Child { get; set; }
-        public int Marriage { get; set; }
-        public int Unpaid { get; set; }
-        public int BloodDonation { get; set; }
-        public int Maternity { get; set; }
-        public int DeathOfRelative { get; set; }
+        public decimal Medical { get; set; }
+        public decimal Yearly { get; set; }
+        public decimal Overtime { get; set; }
+        public decimal Birth { get; set; }
+        public decimal Child { get; set; }
+        public decimal Marriage { get; set; }
+        public decimal Unpaid { get; set; }
+        public decimal BloodDonation { get; set; }
+        public decimal Maternity { get; set; }
+        public decimal DeathOfRelative { get; set; }
     }
 }
